riscv64: Cleanup `ishl` rules

+++ cranelift/codegen/src/isa/riscv64/lower/isle.rs
@@ -213,7 +213,7 @@ impl generated_code::Context for RV64IsleContext<'_, '_, MInst, Riscv64Backend> 
         v.clone()
     }
 
-    fn imm12_and(&mut self, imm: Imm12, x: i32) -> Imm12 {
+    fn imm12_and(&mut self, imm: Imm12, x: u64) -> Imm12 {
         Imm12::from_bits(imm.as_i16() & (x as i16))
     }
